stats/server/requirements.txt
django
djangorestframework
django-filter
mongoengine
numpy
google-cloud
xxhash
shapely
dominate
pyjwt
schedule
